Add stream related javadoc
