+++ pinot-core/src/main/java/com/linkedin/pinot/core/realtime/stream/StreamConfig.java
@@ -19,7 +19,7 @@ import com.google.common.base.Preconditions;
 import com.linkedin.pinot.common.utils.DataSize;
 import com.linkedin.pinot.common.utils.EqualityUtils;
 import com.linkedin.pinot.common.utils.time.TimeUtils;
-import com.linkedin.pinot.core.realtime.impl.kafka.SimpleConsumerFactory;
+import com.linkedin.pinot.core.realtime.impl.kafka.KafkaConsumerFactory;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -45,7 +45,7 @@ public class StreamConfig {
   private static final int DEFAULT_FLUSH_THRESHOLD_ROWS = 5_000_000;
   private static final long DEFAULT_FLUSH_THRESHOLD_TIME = TimeUnit.MILLISECONDS.convert(6, TimeUnit.HOURS);
   private static final long DEFAULT_DESIRED_SEGMENT_SIZE_BYTES = 200 * 1024 * 1024; // 200M
-  private static final String DEFAULT_CONSUMER_FACTORY_CLASS_NAME_STRING = SimpleConsumerFactory.class.getName();
+  private static final String DEFAULT_CONSUMER_FACTORY_CLASS_NAME_STRING = KafkaConsumerFactory.class.getName();
 
   protected static final long DEFAULT_STREAM_CONNECTION_TIMEOUT_MILLIS = 30_000;
   protected static final int DEFAULT_STREAM_FETCH_TIMEOUT_MILLIS = 5_000;
