+++ pinot-integration-tests/src/test/java/com/linkedin/pinot/integration/tests/BaseClusterIntegrationTest.java
@@ -21,7 +21,7 @@ import com.linkedin.pinot.common.config.TableTaskConfig;
 import com.linkedin.pinot.common.utils.KafkaStarterUtils;
 import com.linkedin.pinot.common.utils.TarGzCompressionUtils;
 import com.linkedin.pinot.common.utils.ZkStarter;
-import com.linkedin.pinot.core.realtime.impl.kafka.SimpleConsumerFactory;
+import com.linkedin.pinot.core.realtime.impl.kafka.KafkaConsumerFactory;
 import com.linkedin.pinot.util.TestUtils;
 import java.io.File;
 import java.net.URL;
@@ -100,7 +100,7 @@ public abstract class BaseClusterIntegrationTest extends ClusterTest {
   }
 
   protected String getStreamConsumerFactoryClassName() {
-    return SimpleConsumerFactory.class.getName();
+    return KafkaConsumerFactory.class.getName();
   }
 
   protected int getRealtimeSegmentFlushSize() {
